fit content to page and make the table and header as components
