+++ src/app/dashboard/page.tsx
@@ -1,74 +1,14 @@
 'use client'
 
 import React from "react";
-
-import {
-  Table,
-  TableHeader,
-  TableBody,
-  TableColumn,
-  TableRow,
-  TableCell,
-  getKeyValue,
-
-  Navbar, 
-  NavbarBrand, 
-  NavbarContent, 
-  NavbarItem, 
-  NavbarMenuToggle,
-  NavbarMenu,
-  NavbarMenuItem,
-  Link,
-  Button
-} from "@nextui-org/react";
+import Header from "@/components/Header"
+import Expences from "@/components/Expences"
 
 export default function Dashboard() {
-  const rows = [
-    {
-      key: "1",
-      category: "rent",
-      payment: "rent",
-      amount: -950,
-    },
-    {
-      key: "2",
-      category: "rent",
-      payment: "rent",
-      amount: -950,
-    },
-    {
-      key: "3",
-      category: "rent",
-      payment: "rent",
-      amount: -950,
-    },
-  ];
-
-  const columns = [
-    {
-      key: "category",
-      label: "Category",
-    },
-    {
-      key: "payment",
-      label: "Payment",
-    },
-    {
-      key: "amount",
-      label: "Amount",
-    },
-  ];
-
   return (
-    <div className="dark:bg-teal-800" >
-      <Navbar>
-        <NavbarBrand className="font-bold text-2xl" >Finesse</NavbarBrand>
-        <NavbarContent justify="end">
-          <NavbarItem><Link color="foreground" href="#">Login</Link></NavbarItem>
-          <NavbarItem><Link color="foreground" href="#">Sign Up</Link></NavbarItem>
-        </NavbarContent>
-      </Navbar>
-      <main className="flex justify-evenly">
+    <div className="dark:bg-teal-800 flex flex-col min-h-screen">
+      <Header/>
+      <main className="flex justify-evenly flex-grow">
         <div>
           <div className="flex justify-evenly">
             <div>Bar</div>
@@ -77,28 +17,13 @@ export default function Dashboard() {
               <p><b>Expences</b> {"-950"}</p>
             </div>
           </div>
-          <Table>
-            <TableHeader>
-              {columns.map((column) => (
-                <TableColumn key={column.key}>{column.label}</TableColumn>
-              ))}
-            </TableHeader>
-            <TableBody>
-              {rows.map((row) => (
-                <TableRow key={row.key}>
-                  {(columnKey) => (
-                    <TableCell className="font-white">{getKeyValue(row, columnKey)}</TableCell>
-                  )}
-                </TableRow>
-              ))}
-            </TableBody>
-          </Table>
+          <Expences/>
         </div>
         <div>
           <p>Chart</p>
         </div>
       </main>
-      <footer className="flex justify-center">footer</footer>
+      <footer className="flex justify-center h-20 bg-gray-700">footer</footer>
     </div>
   );
 }
